requirements.txt
asgiref==3.4.0
Django==3.2.4
django-phone-field==1.8.1
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
